refactor(review-card): ♻️ file organization and implement in related pages
