+++ src/lib/components/review_card/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewCard } from './card.svelte';
