+++ src/lib/components/review_card/racket_fields.svelte
