+++ src/lib/components/review_card/string_fields.svelte
